+++ dandy/llm/service/service.py
@@ -14,8 +14,9 @@ from dandy.intel.factory import IntelFactory
 from dandy.intel.type_vars import IntelType
 from dandy.llm.exceptions import LlmCriticalException, LlmValidationCriticalException, LlmRecoverableException
 from dandy.llm.prompt import Prompt
-from dandy.llm.service.recorder import recorder_add_llm_request_event, recorder_add_llm_response_event, recorder_add_llm_success_event, \
-    recorder_add_llm_validation_failure_event, recorder_add_llm_retry_event
+from dandy.llm.service.recorder import recorder_add_llm_request_event, recorder_add_llm_response_event, \
+    recorder_add_llm_success_event, \
+    recorder_add_llm_failure_event, recorder_add_llm_retry_event
 from dandy.llm.service.prompts import service_system_validation_error_prompt, service_user_prompt, \
     service_system_prompt
 
@@ -34,16 +35,24 @@ class LlmService(BaseHttpService):
     ):
         super().__init__(config=llm_config.http_config)
 
+        self.event_id = generate_new_recorder_event_id()
+
         self._llm_config = llm_config
         self._llm_options = llm_options
-
-    def get_request_body(self) -> BaseRequestBody:
-        return self._llm_config.generate_request_body(
+        self._intel = None
+        self._intel_json_schema = None
+        self._request_body = self._llm_config.generate_request_body(
             max_input_tokens=self._llm_options.max_input_tokens,
             max_output_tokens=self._llm_options.max_output_tokens,
             seed=self._llm_options.seed,
             temperature=self._llm_options.temperature,
         )
+        self._response_content = None
+        self._retry_attempt = 0
+
+    @property
+    def has_retry_attempts_available(self) -> bool:
+        return self._retry_attempt < self._llm_config.options.prompt_retry_count
 
     def process_prompt_to_intel(
             self,
@@ -63,21 +72,19 @@ class LlmService(BaseHttpService):
         if intel_class is None and intel_object is None:
             raise LlmCriticalException('Must specify either intel_class or intel_object.')
 
-        intel_json_schema = IntelFactory.intel_to_json_inc_ex_schema(
-            intel=intel_class or intel_object,
+        self._intel = intel_class or intel_object
+
+        self._intel_json_schema = IntelFactory.intel_to_json_inc_ex_schema(
+            intel=self._intel,
             include=include_fields,
             exclude=exclude_fields
         )
 
-        event_id = generate_new_recorder_event_id()
-
-        request_body = self.get_request_body()
-
-        request_body.set_format_to_json_schema(
-            intel_json_schema
+        self._request_body.set_format_to_json_schema(
+            self._intel_json_schema
         )
 
-        request_body.add_message(
+        self._request_body.add_message(
             role='system',
             content=service_system_prompt(
                 system_prompt=system_prompt
@@ -86,13 +93,13 @@ class LlmService(BaseHttpService):
 
         if message_history:
             for message in message_history.messages:
-                request_body.add_message(
+                self._request_body.add_message(
                     role=message.role,
                     content=message.content,
                     images=message.images
                 )
 
-        request_body.add_message(
+        self._request_body.add_message(
             role='user',
             content=service_user_prompt(
                 prompt if isinstance(prompt, Prompt) else Prompt(prompt)
@@ -100,42 +107,31 @@ class LlmService(BaseHttpService):
             images=images,
         )
 
-        return self._process_request_to_intel(
-            request_body=request_body,
-            intel_json_schema=intel_json_schema,
-            event_id=event_id,
-            intel=intel_class or intel_object
-        )
-
+        return self._process_request_to_intel()
 
     def _process_request_to_intel(
             self,
-            request_body: BaseRequestBody,
-            intel_json_schema: dict,
-            event_id: str,
-            intel: Union[BaseIntel, Type[BaseIntel]],
-            retry_attempt: int = 0,
     ) -> IntelType:
-        recorder_add_llm_request_event(request_body, intel_json_schema, event_id)
+        recorder_add_llm_request_event(self._request_body, self._intel_json_schema, self.event_id)
 
-        message_content = self._llm_config.get_response_content(
+        self._response_content = self._llm_config.get_response_content(
             self.post_request(
-                request_body.model_dump()
+                self._request_body.model_dump()
             )
         )
 
-        recorder_add_llm_response_event(message_content, event_id)
+        recorder_add_llm_response_event(self._response_content, self.event_id)
 
         try:
             intel_object = IntelFactory.json_to_intel_object(
-                message_content,
-                intel
+                self._response_content,
+                self._intel
             )
 
             if intel_object is not None:
                 recorder_add_llm_success_event(
                     'Validated response from prompt into intel object.',
-                    event_id,
+                    self.event_id,
                     intel=intel_object
                 )
 
@@ -144,37 +140,41 @@ class LlmService(BaseHttpService):
             else:
                 raise LlmRecoverableException('Failed to validate response from prompt into intel object.')
 
-        # This validation error would be raised by pydantic if the IntelFactory failed to parse the JSON into an Intel object
-        except ValidationError as e:
-            recorder_add_llm_validation_failure_event(e, event_id)
+        except ValidationError as error:
+            recorder_add_llm_failure_event(error, self.event_id)
 
-            if retry_attempt < self._llm_config.options.prompt_retry_count:
-                recorder_add_llm_retry_event(
-                    'Validation of response to intel object failed retrying with validation errors prompt.',
-                    event_id,
-                    remaining_attempts = self._llm_config.options.prompt_retry_count - (retry_attempt + 1)
-                )
+            return self.retry_process_request_to_intel(
+                retry_event_description='Validation of response to intel object failed, retrying with validation errors prompt.',
+                retry_user_prompt=service_system_validation_error_prompt(error)
+            )
 
-                request_body.add_message(
-                    role='assistant',
-                    content=message_content
-                )
+    def retry_process_request_to_intel(
+            self,
+            retry_event_description: str,
+            retry_user_prompt: Prompt | str,
+    ) -> IntelType:
+        if self.has_retry_attempts_available:
+            self._retry_attempt += 1
 
-                request_body.add_message(
-                    role='user',
-                    content=service_system_validation_error_prompt(e).to_str()
-                )
+            recorder_add_llm_retry_event(
+                retry_event_description,
+                self.event_id,
+                remaining_attempts=self._llm_config.options.prompt_retry_count - self._retry_attempt
+            )
 
-                return self._process_request_to_intel(
-                    request_body=request_body,
-                    intel_json_schema=intel_json_schema,
-                    event_id=event_id,
-                    intel=intel,
-                    retry_attempt=retry_attempt + 1
-                )
+            self._request_body.add_message(
+                role='assistant',
+                content=self._response_content
+            )
 
-            else:
-                raise LlmRecoverableException(f'Failed to validate response from prompt into intel object after {self._llm_config.options.prompt_retry_count} attempts.')
+            self._request_body.add_message(
+                role='user',
+                content=Prompt(retry_user_prompt).to_str()
+            )
 
+            return self._process_request_to_intel(
+            )
 
-
+        else:
+            raise LlmRecoverableException(
+                f'Failed to get the correct response from the LlmService after {self._llm_config.options.prompt_retry_count} attempts.')